requirements.txt
google-cloud-bigquery==3.1.0
google-cloud-storage==1.42.0
requests==2.26.0
pyarrow==7.0.0
pandas==1.3.3
python-dateutil==2.8.2
gcsfs==2023.10.0
shapely<2.0